refactor(useTeams): extract getTeams function into seperate util function
